Working on adding Users to specific privileges/roles. (WIP) Adds SetupDataLoader, UserDto, IUserService, Privilege, PrivilegeRepository, UserAlreadyExistsException

+++ h2-demo/src/main/java/com/teamsparq/h2demo/payload/response/UserInfoResponse.java
@@ -3,26 +3,23 @@ package com.teamsparq.h2demo.payload.response;
 import java.util.List;
 
 public class UserInfoResponse {
-    private Integer id;
+    private Long id;
     private String username;
     private String email;
-
-
-
     private List<String> roles;
 
-    public UserInfoResponse(Integer id, String username, String email, List<String> roles) {
+    public UserInfoResponse(Long id, String username, String email, List<String> roles) {
         this.id = id;
         this.username = username;
         this.email = email;
         this.roles = roles;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
